GPX handling & homepage cleanup & map
GPX routes are now shown in a map on the homepage, which has been placed
in the top left hand corner of the screen. The map and the album cover
tiles now occupy a html grid as well, which expands based on screen
resolution.

There have been no db changes, a simple server build is all that is
required to perform an update. Note that the client requests a gpx file
from every album - albums with no gpx file will cause 404s. It is also
possible to change map styles, simply by updating the 'topo.json' file
in ```static/```.